requirements.txt
rich
watchdog
editor
GitPython